requirements.txt
google-generativeai==0.5.2
matplotlib==3.8.4
streamlit==1.29.0
yfinance==0.2.38